@@ -12,17 +12,16 @@
     }
   ],
   "require": {
-    "php": "^8.0",
+    "php": ">=8.1",
     "ext-bcmath": "*"
   },
   "require-dev": {
-    "friendsofphp/php-cs-fixer": "^3.13",
-    "phpstan/phpstan": "^1.9",
-    "phpstan/phpstan-phpunit": "^1.3",
-    "phpstan/phpstan-webmozart-assert": "^1.2",
-    "phpunit/phpunit": "^9.5",
+    "friendsofphp/php-cs-fixer": "^3.51",
+    "infection/infection": "^0.27.10",
+    "phpstan/extension-installer": "*",
+    "phpunit/phpunit": "^10.5 | ^11.0",
     "psalm/plugin-phpunit": "^0.18",
-    "squizlabs/php_codesniffer": "^3.7",
+    "squizlabs/php_codesniffer": "^3.9",
     "vimeo/psalm": "^5.4"
   },
   "autoload": {
@@ -32,16 +31,21 @@
     }
   },
   "config": {
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "infection/extension-installer": true,
+      "phpstan/extension-installer": true
+    }
   },
   "prefer-stable": true,
   "scripts": {
     "phpcs": "phpcs --standard=phpcs.xml.dist",
     "cs-check": "php-cs-fixer -v --dry-run --using-cache=no fix",
     "cs-fix": "php-cs-fixer --using-cache=no fix",
-    "test": "export XDEBUG_MODE=coverage && vendor/bin/phpunit --configuration phpunit.xml",
+    "test": "export XDEBUG_MODE=coverage && phpunit -c phpunit.xml --coverage-xml build/logs --coverage-clover build/logs/clover.xml --coverage-html build/logs/clover.html --log-junit build/logs/junit.xml",
     "psalm": "vendor/bin/psalm --no-cache",
-    "analyze": "vendor/bin/phpstan analyse --xdebug --configuration phpstan.neon.dist",
+    "infection": "infection --threads=4",
+    "infection-ci": "infection --coverage=build/logs --threads=4",
     "check": [
       "@phpcs",
       "@cs-check",
