@@ -4,31 +4,30 @@ declare(strict_types=1);
 
 namespace MarcelStrahl\PriceCalculator\Tests\Helpers\Converter\Currencies;
 
-use MarcelStrahl\PriceCalculator\Helpers\Converter\ConverterInterface;
+use MarcelStrahl\PriceCalculator\Contracts\Converter\ConverterInterface;
 use MarcelStrahl\PriceCalculator\Helpers\Converter\Currencies\EuroToCent;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @author Marcel Strahl <info@marcel-strahl.de>
  */
-class EuroToCentTest extends TestCase
+#[CoversClass(className: EuroToCent::class)]
+final class EuroToCentTest extends TestCase
 {
-    /**
-     * @return void
-     */
-    public function testImplements(): void
+    #[Test]
+    public function hasImplemented(): void
     {
         $converter = new EuroToCent();
         $this->assertInstanceOf(ConverterInterface::class, $converter);
         $this->assertInstanceOf(EuroToCent::class, $converter);
     }
 
-    /**
-     * @dataProvider dataProviderConvert
-     * @param float $amount
-     * @param float|null $expected
-     */
-    public function testConvert(float $amount, ?float $expected): void
+    #[Test]
+    #[DataProvider(methodName: 'dataProviderConvert')]
+    public function canConvert(float $amount, ?float $expected): void
     {
         $converter = new EuroToCent();
 
@@ -36,9 +35,14 @@ class EuroToCentTest extends TestCase
     }
 
     /**
-     * @return array
+     * @return array{
+     *    0: array{0: 1.00, 1: 100.00},
+     *    1: array{0: 1.15, 1: 115.00},
+     *    2: array{0: 116.80, 1: 11680.00},
+     *    3: array{0: .00, 1: 0}
+     * }
      */
-    public function dataProviderConvert(): array
+    public static function dataProviderConvert(): array
     {
         return [
             [1.00, 100.00],
