@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace MarcelStrahl\PriceCalculator\Facade;
 
+use MarcelStrahl\PriceCalculator\Contracts\PriceCalculatorInterface;
 use MarcelStrahl\PriceCalculator\PriceCalculator as PriceCalculatorService;
-use MarcelStrahl\PriceCalculator\PriceCalculatorInterface;
 
 /**
  * @author Marcel Strahl <info@marcel-strahl.de>
