@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace MarcelStrahl\PriceCalculator\Facade;
 
+use MarcelStrahl\PriceCalculator\Contracts\PriceCalculatorInterface;
 use MarcelStrahl\PriceCalculator\Helpers\Entity\Vat;
-use MarcelStrahl\PriceCalculator\PriceCalculatorInterface;
 use MarcelStrahl\PriceCalculator\Service\VatCalculator as VatCalculatorService;
 
 /**
@@ -13,7 +13,9 @@ use MarcelStrahl\PriceCalculator\Service\VatCalculator as VatCalculatorService;
  */
 class VatCalculator
 {
-    private function __construct(private Vat $vat, private PriceCalculatorInterface $priceCalculator) {}
+    private function __construct(private Vat $vat, private PriceCalculatorInterface $priceCalculator)
+    {
+    }
 
     public static function getVatCalculator(int $vat): VatCalculatorService
     {
