@@ -4,8 +4,9 @@ declare(strict_types=1);
 
 namespace MarcelStrahl\PriceCalculator\Factory;
 
+use MarcelStrahl\PriceCalculator\Contracts\Converter\ConverterInterface;
+use MarcelStrahl\PriceCalculator\Contracts\Factory\ConverterFactoryInterface;
 use MarcelStrahl\PriceCalculator\Exceptions\PriceCalculatorFactoryException;
-use MarcelStrahl\PriceCalculator\Helpers\Converter\ConverterInterface;
 use MarcelStrahl\PriceCalculator\Helpers\Converter\Currencies\CentToEuro;
 use MarcelStrahl\PriceCalculator\Helpers\Converter\Currencies\EuroToCent;
 
