@@ -4,7 +4,8 @@ declare(strict_types=1);
 
 namespace MarcelStrahl\PriceCalculator\Helpers\Converter\Currencies;
 
-use MarcelStrahl\PriceCalculator\Helpers\Converter\ConverterInterface;
+use MarcelStrahl\PriceCalculator\Contracts\Converter\ConverterInterface;
+use MarcelStrahl\PriceCalculator\Contracts\Type\FiguresInterface;
 
 /**
  * @author Marcel Strahl <info@marcel-strahl.de>
@@ -17,10 +18,10 @@ class CentToEuro implements ConverterInterface
     public function convert(float $amount): float
     {
         if ($this->isEmpty($amount)) {
-            return .0;
+            return FiguresInterface::FLOAT_ZERO;
         }
 
-        return (float) bcdiv((string) $amount, '100', 2);
+        return (float) bcdiv((string) $amount, FiguresInterface::STRING_HUNDRED, FiguresInterface::INTEGER_TWO);
     }
 
     /**
@@ -28,6 +29,6 @@ class CentToEuro implements ConverterInterface
      */
     public function isEmpty(float $amount): bool
     {
-        return empty($amount) || $amount < 0;
+        return empty($amount);
     }
 }
