@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 namespace MarcelStrahl\PriceCalculator\Helpers\Entity;
 
-use MarcelStrahl\PriceCalculator\Helpers\Types\DiscountInterface;
+use MarcelStrahl\PriceCalculator\Contracts\Type\DiscountInterface;
 
 /**
  * @author Marcel Strahl <info@marcel-strahl.de>
