@@ -4,28 +4,27 @@ declare(strict_types=1);
 
 namespace MarcelStrahl\PriceCalculator\Tests\Facade;
 
+use MarcelStrahl\PriceCalculator\Contracts\PriceCalculatorInterface;
 use MarcelStrahl\PriceCalculator\Facade\PriceCalculator as PriceCalculatorFacade;
-use MarcelStrahl\PriceCalculator\PriceCalculatorInterface;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 
 /**
  * @author Marcel Strahl <info@marcel-strahl.de>
  */
-class PriceCalculatorTest extends TestCase
+#[CoversClass(className: PriceCalculatorFacade::class)]
+final class PriceCalculatorTest extends TestCase
 {
-    /**
-     * @return void
-     */
-    public function testCanInitPriceCalculatorFacade(): void
+    #[Test]
+    public function canInitPriceCalculatorFacade(): void
     {
         $facade = new PriceCalculatorFacade();
         $this->assertInstanceOf(PriceCalculatorFacade::class, $facade);
     }
 
-    /**
-     * @return void
-     */
-    public function testGetPriceCalculator(): void
+    #[Test]
+    public function canGetPriceCalculator(): void
     {
         $facade = new PriceCalculatorFacade();
         $priceCalculator = $facade::getPriceCalculator();
